Move Message types to own class
